GenericTypeReflector: replace stream array-mapping with custom helper
Streams is a very flexible way to process data, but it involves a lot of
classes and complexity. In this case we are always doing the exact
same fixed size array mapping, so we don't need all the features
that comes with Streams, and can just use an efficient helper.